backend/requirements.txt
fastapi
uvicorn[standard]
httpx
supabase
python-dotenv
pydantic
requests
bcrypt
python-jose[cryptography]
pytest
pytest-asyncio
pytest-asyncio

